Removal of AzQtAppWindow
