--- Code/Framework/AzQtComponents/AzQtComponents/Application/Window/AzQtApplicationWindow.cpp
@@ -1,29 +1,0 @@
-/*
-* All or portions of this file Copyright (c) Amazon.com, Inc. or its affiliates or
-* its licensors.
-*
-* For complete copyright and license terms please see the LICENSE at the root of this
-* distribution (the "License"). All use of this software is governed by the License,
-* or, if provided, by the license below or the license accompanying this file. Do not
-* remove or modify any license notices. This file is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*
-*/
-
-#include <AzQtComponents/Application/Window/AzQtApplicationWindow.h>
-#include <QVBoxLayout>
-
-
-namespace AzQtComponents
-{
-    AzQtApplicationWindow::AzQtApplicationWindow(QWidget* parent /* = 0 */)
-        : AzQtComponents::DockMainWindow(parent)
-    {
-        m_advancedDockManager = new AzQtComponents::FancyDocking(this);
-
-        m_menuBar = new QMenuBar(this);
-        m_menuBar->setObjectName("MenuBar");
-        setMenuBar(m_menuBar);
-    }
-}
-
