+++ Gems/Atom/Tools/MaterialEditor/Code/Source/Window/MaterialEditorWindow.h
@@ -12,11 +12,10 @@
 #include <Atom/Document/MaterialDocumentNotificationBus.h>
 
 AZ_PUSH_DISABLE_WARNING(4251 4800, "-Wunknown-warning-option") // disable warnings spawned by QT
-//#include <AzQtComponents/Components/DockMainWindow.h>
+#include <AzQtComponents/Components/DockMainWindow.h>
 #include <AzQtComponents/Components/FancyDocking.h>
 #include <AzQtComponents/Components/StyledDockWidget.h>
 #include <AzQtComponents/Components/Widgets/TabWidget.h>
-#include <AzQtComponents/Application/Window/AzQtApplicationWindow.h>
 
 #include <Atom/Window/MaterialEditorWindowRequestBus.h>
 #include <Viewport/MaterialViewportWidget.h>
@@ -44,7 +43,7 @@ namespace MaterialEditor
      * 3) MaterialPropertyInspector  - The user edits the properties of the selected Material.
      */
     class MaterialEditorWindow
-        : public AzQtComponents::AzQtApplicationWindow
+        : public AzQtComponents::DockMainWindow
         , private MaterialEditorWindowRequestBus::Handler
         , private MaterialDocumentNotificationBus::Handler
     {
@@ -75,8 +74,8 @@ namespace MaterialEditor
         void OnDocumentUndoStateChanged(const AZ::Uuid& documentId) override;
         void OnDocumentSaved(const AZ::Uuid& documentId) override;
 
-        void SetupMenu() override;
-        void SetupTabs() override;
+        void SetupMenu();
+        void SetupTabs();
 
         void AddTabForDocumentId(const AZ::Uuid& documentId);
         void RemoveTabForDocumentId(const AZ::Uuid& documentId);
@@ -84,12 +83,14 @@ namespace MaterialEditor
         AZ::Uuid GetDocumentIdFromTab(const int tabIndex) const;
         QString GetDocumentPath(const AZ::Uuid& documentId) const;
 
-        void OpenTabContextMenu() override;
+        void OpenTabContextMenu();
         void SelectPreviousTab();
         void SelectNextTab();
 
         void closeEvent(QCloseEvent* closeEvent) override;
 
+        AzQtComponents::FancyDocking* m_advancedDockManager = nullptr;
+        QMenuBar* m_menuBar = nullptr;
         QWidget* m_centralWidget = nullptr;
         AzQtComponents::TabWidget* m_tabWidget = nullptr;
         MaterialViewportWidget* m_materialViewport = nullptr;
