+++ Gems/Atom/Tools/ShaderManagementConsole/Code/Source/Window/ShaderManagementConsoleWindow.h
@@ -15,11 +15,10 @@
 #include <Atom/RPI.Edit/Shader/ShaderVariantListSourceData.h>
 
 AZ_PUSH_DISABLE_WARNING(4251 4800, "-Wunknown-warning-option") // disable warnings spawned by QT
-//#include <AzQtComponents/Components/DockMainWindow.h>
+#include <AzQtComponents/Components/DockMainWindow.h>
 #include <AzQtComponents/Components/FancyDocking.h>
 #include <AzQtComponents/Components/StyledDockWidget.h>
 #include <AzQtComponents/Components/Widgets/TabWidget.h>
-#include <AzQtComponents/Application/Window/AzQtApplicationWindow.h>
 
 #include <Window/ShaderManagementConsoleBrowserWidget.h>
 #include <Window/ToolBar/ShaderManagementConsoleToolBar.h>
@@ -42,7 +41,7 @@ namespace ShaderManagementConsole
      * its panels, managing selection of assets, and performing high-level actions like saving. It contains...
      */
     class ShaderManagementConsoleWindow
-        : public AzQtComponents::AzQtApplicationWindow
+        : public AzQtComponents::DockMainWindow
         , private ShaderManagementConsoleDocumentNotificationBus::Handler
     {
         Q_OBJECT
@@ -60,15 +59,15 @@ namespace ShaderManagementConsole
         void OnDocumentUndoStateChanged(const AZ::Uuid& documentId) override;
         void OnDocumentSaved(const AZ::Uuid& documentId) override;
 
-        void SetupMenu() override;
-        void SetupTabs() override;
+        void SetupMenu();
+        void SetupTabs();
 
         void AddTabForDocumentId(const AZ::Uuid& documentId);
         void RemoveTabForDocumentId(const AZ::Uuid& documentId);
         void UpdateTabForDocumentId(const AZ::Uuid& documentId);
         AZ::Uuid GetDocumentIdFromTab(const int tabIndex) const;
 
-        void OpenTabContextMenu() override;
+        void OpenTabContextMenu();
         void SelectPreviousTab();
         void SelectNextTab();
 
@@ -80,6 +79,8 @@ namespace ShaderManagementConsole
 
         void CreateDocumentContent(const AZ::Uuid& documentId, QStandardItemModel* model);
 
+        AzQtComponents::FancyDocking* m_advancedDockManager = nullptr;
+        QMenuBar* m_menuBar = nullptr;
         QWidget* m_centralWidget = nullptr;
         AzQtComponents::TabWidget* m_tabWidget = nullptr;
         ShaderManagementConsoleBrowserWidget* m_assetBrowser = nullptr;
